requirements.txt
streamlit==1.29.0
google-generativeai>=0.4.1,<0.5.0
langchain==0.1.5
langchain-google-genai==0.0.11
PyPDF2==3.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0